fix(rocjitsu): make the snapshot entry's debug_prop narrowing explicit
DebugTopology::debug_prop is uint64_t because that is the width of the field's
canonical home: the kernel's node property is 64-bit (HSA_DBG_RESERVED is
0xfffffffffffff000ull) and libhsakmt parses it into an HSAuint64 bitfield union
with a 52-bit Reserved tail. The DBG_TRAP snapshot entry is the outlier --
kfd_dbg_device_info_entry::debug_prop is __u32 -- and real libhsakmt narrows the
same way without a cast (debug.c, hsaKmtDbgGetDeviceDataCtx).

The build carries no -Wconversion, so the assignment truncated silently. Every
bit the derivation sets lives below 0x1000, so nothing real is lost today, but a
config-supplied value above 2^32 would have sysfs publish the full 64 bits while
the snapshot reported the low 32 -- the two-paths-disagree divergence this area
exists to prevent.

Spell the narrowing out and record why the widths differ.

+++ emulation/rocjitsu/lib/rocjitsu/src/rocjitsu/kmd/linux/simulated_kfd.cpp
@@ -2277,7 +2277,11 @@ int SimulatedKfd::debug_device_snapshot(kfd_ioctl_dbg_trap_device_snapshot_args 
         kmd::effective_topology_for(info.gfx_target_version, info.capability, info.capability2,
                                     info.debug_prop, info.revision_id);
     e.capability = topology.capability;
-    e.debug_prop = topology.debug_prop;
+    // debug_prop is __u32 in the snapshot entry but __u64 in the sysfs node
+    // property, so a config that captured a debug_prop above 2^32 would have
+    // the two paths report different values. The derived bits all fit; make the
+    // narrowing the uapi struct imposes explicit rather than incidental.
+    e.debug_prop = static_cast<uint32_t>(topology.debug_prop);
 
     std::memcpy(out + static_cast<uint64_t>(i) * in_entry_size, &e, args.entry_size);
   }
